chore: move MTableBodyRow to own directory
