+++ src/components/index.js
@@ -13,7 +13,6 @@ export { default as Container } from './Container';
 
 /** Still needs to be refactored into functional */
 export { default as MTableBody } from './m-table-body';
-export { default as MTableBodyRow } from './m-table-body-row';
 /** HAVING ISSUES WITH THE REFACTORED VERSIONS OF: */
 export { default as MTableEditField } from './m-table-edit-field';
 export { default as MTableEditCell } from './m-table-edit-cell';
@@ -26,6 +25,7 @@ export { default as MTableEditCell } from './m-table-edit-cell';
 // Trying to keep these in alphabetical order
 export { default as MTableAction } from './MTableAction';
 export { default as MTableActions } from './MTableActions';
+export { default as MTableBodyRow } from './MTableBodyRow.js';
 export { default as MTableCell } from './MTableCell';
 export { default as MTableEditRow } from './MTableEditRow';
 export { default as MTableFilterRow } from './MTableFilterRow';
@@ -34,7 +34,7 @@ export { default as MTableGroupRow } from './MTableGroupRow';
 export { default as MTableHeader } from './MTableHeader';
 export { default as MTableSteppedPagination } from './MTableSteppedPaginationInner';
 export { default as MTablePagination } from './MTablePagination';
-export { default as MTableSummaryRow } from "./MTableSummaryRow";
+export { default as MTableSummaryRow } from './MTableSummaryRow';
 export { default as MTableToolbar } from './MTableToolbar';
 /** THESE REFACTORS ARE HAVING ISSUES */
 // export { default as MTableEditCell } from './MTableEditCell';
